test/fixtures/wpt/wasm/webapi/esm-integration/execute-start.tentative.html

<!DOCTYPE html>
<title>Check execution of WebAssembly start function</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    setup({allow_uncaught_exception: true});

    const test_load = async_test(
        "Importing a WebAssembly module should execute the start function.");

    window.log = [];
    window.addEventListener("error", ev => {
      log.push(ev.message);
    });

    window.addEventListener("load", test_load.step_func_done(ev => {
      assert_array_equals(log, ["executed"]);
    }));

    function unreachable() { log.push("unexpected"); }
</script>
<script type="module" src="./resources/execute-start.wasm" onerror="unreachable()""></script>
